Move FeePerKbStorageIndexKey into feeperkb package
